requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
pytesseract>=0.3.10
pillow>=10.0.0
ollama>=0.2.0
numpy>=1.26.0
